Update bio.json
Your MIS portfolio does not work in the MIS Gallary because your JSON bio had 3 extra commas in various places. The last item in a JSON array should not end with a comma. See the edits I made for details. Merge this pull request if it seems ok and let's see if your portfolio works in the MIS gallery.